backend/requirements.txt
pydantic==1.10.13 
fastapi==0.100.0   # Use older version
uvicorn[standard]==0.23.0 
python-dotenv==1.0.0 
google-generativeai==0.3.0 
